requirements.txt
aniso8601==7.0.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
coverage==7.6.1
Django==5.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-filter==24.3
django-graphql-auth==0.3.16
django-js-asset==2.2.0
django-mptt==0.16.0
django-nine==0.2.7
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.2
factory_boy==3.3.1
Faker==27.0.0
flake8==7.1.1
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
promise==2.3
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==1.7.1
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
rpds-py==0.20.0
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
sqlparse==0.5.1
standard-distutils==3.9.19
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
